% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_orders}
\alias{event_orders}
\title{Orders from an event venue.}
\format{A tbl_df with 20698 rows an 10 variables:
\describe{
  \item{uid}{unique identifier for customer (integer)}
  \item{order_on}{order date (Date)}
  \item{num_seats}{# seats in order (integer)}
  \item{value}{total dollar value for order (numeric)}
}}
\source{
sanitized data from a DS4CI customer
}
\usage{
event_orders
}
\description{
20698 orders from 5345 customers over ~23 months from 2014-15.
}
\keyword{datasets}
